feat:db-init-status finished

+++ Backend/ticketing-server/src/main/java/com/ticketing/queue/DTO/QueueUserInfoDTO.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class QueueUserInfoDTO {
     int clickMiss;
-    int duration;
+    float duration;
 }
